Add initial parsing support for hierarchical PDs
This change allows PDs to be defined inside a parent PD XML tag,
providing a tree like data structure.

At this point the tree is flattened and then passed to the system
generator. So:

    <protection_domain name="a" ...>
        <protection_domain name="b" ... />
        <protection_domain name="c" ... />
    </protection_domain>

Is exactly the same as:

    <protection_domain name="a" ... />
    <protection_domain name="b" ... />
    <protection_domain name="c" ... />

At this point in time. This provides the basis for the next step
of adding fault handling.