libsel4cp: load from libdummycrt
This allows libsel4cp to use to use TLS and be loaded exactly the same
way as threaded PDs in the future. This commit also removes the _stack
static allocation, and leaving the allocation of the stack to the sel4cp
tool instead.

In the future, libsel4cp will be renamed to something like
libevents, providing an event-driven model for PDs to use.