tool: fix off-by-x error with child thread objects
Currently, if we have N max threads, N - (# child PDs) will be the
max spawnable threads for a root PD. However, we allocate the child
thread objects based on N and not the actual amount is used.

Therefore, when we refer to the FIRST spawnable thread object, it is
at (# child PDs) and not 0, as was previously used.

This means that if N=10, we have the following layout in the root
CNode for child threads, where A = unused and B is a spawnable thread.
Take (# child PDs) = 3.

     Spawnable objects start here
     \/
A A A B B B B B B B
^
| 3 unused objects

This is obviously wasteful, and we can come up with a better scheme.
However, it is important that in the future we account for this in
both the code and the tool.