tool: change the way root PPCs are handled
Previously, to create a root PPC, we specified a channel from the child
PD to the root PD. This is by itself a good idea, it makes use of
existing mechanisms in sel4cp and doesn't make things complicated.
However, with the support of _threading_, each thread needs its own
endpoint to communicate on. Either this means messing with the current
channel abstraction such that each channel can have different threads of
origin, or we can just hardcode an endpoint for each thread if we know
the root PD is taking root PPCs.