sdk: clean up thread creation
Let the thread entry in the C runtime setup the stack, IPC buffer and
TLS. We just need to pass in the thread ID. This simplifies the amount
of work the OS has to do. The OS API layer, however, has to be aware
that a thread's starting point should always be at sel4cp_thread_entry.