libsel4cp: clarify sel4cp_thread_set_priority arg1
We use another thread's TCB as the authority not a PD's. In the future,
we want to be able to specify the root PD's thread as well, this will be
useful for when we want the priority to be higher than ANY of the child
thread's priorities (e.g. Error Handler).